REFACTOR: change time type

+++ schedule/core/src/main/kotlin/vp/togedo/data/dao/FixedScheduleDao.kt
@@ -4,8 +4,8 @@ import org.bson.types.ObjectId
 
 data class FixedScheduleDao(
     var scheduleId: ObjectId?,
-    val startTime: Int,
-    val endTime: Int,
+    val startTime: Long,
+    val endTime: Long,
     val title: String,
     val color: String
 )
