+++ schedule/presentation/src/main/kotlin/vp/togedo/data/dto/flexiblePersonalSchedule/CreateFlexibleReqDto.kt
@@ -1,8 +1,8 @@
 package vp.togedo.data.dto.flexiblePersonalSchedule
 
 data class CreateFlexibleReqDto(
-    val startTime: Int,
-    val endTime: Int,
+    val startTime: Long,
+    val endTime: Long,
     val title: String,
     val color: String
 )
