+++ schedule/presentation/src/main/kotlin/vp/togedo/data/dto/flexiblePersonalSchedule/UpdateFlexibleReqDto.kt
@@ -3,8 +3,8 @@ package vp.togedo.data.dto.flexiblePersonalSchedule
 
 data class UpdateFlexibleReqDto(
     val id: String,
-    val startTime: Int,
-    val endTime: Int,
+    val startTime: Long,
+    val endTime: Long,
     val title: String,
     val color: String
 )
